requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0

# ── Database ─────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# ── Validation & Settings ───────────────────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# ── Authentication ───────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18

# ── Email ────────────────────────────────────────────────────────────────
aiosmtplib>=3.0.0
jinja2>=3.1.5

# ── HTTP Client (Google/Meta API calls) ──────────────────────────────────
httpx>=0.28.0

# ── Environment ──────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Google API ───────────────────────────────────────────────────────────
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.160.0
